Add manual-specific instructions; fix Python link
As written, this document doesn't provide any specific instructions for the Csound manual. Instead, the instructions are generic. I found that very confusing, so I have added an introduction that provides Csound-manual-specific instructions. I have also updated the reference to the Python installer. As written, the document links to outdated instructions for installing Python 2. Rather than linking to instructions for some specific version of Python, it makes more sense to link to the Python downloads page. The user will then always be able to find instructions for installing or building the current (or some other) version of Python.